+++ sentry-ruby/spec/sentry/breadcrumb_spec.rb
@@ -30,6 +30,25 @@ RSpec.describe Sentry::Breadcrumb do
     )
   end
 
+  describe "#initialize" do
+    it "limits the maximum size of message" do
+      long_message = "a" * Sentry::Event::MAX_MESSAGE_SIZE_IN_BYTES * 2
+
+      crumb = described_class.new(message: long_message)
+      expect(crumb.message.length).to eq(Sentry::Event::MAX_MESSAGE_SIZE_IN_BYTES + 1)
+    end
+  end
+
+  describe "#message=" do
+    it "limits the maximum size of message" do
+      long_message = "a" * Sentry::Event::MAX_MESSAGE_SIZE_IN_BYTES * 2
+
+      crumb = described_class.new
+      crumb.message = long_message
+      expect(crumb.message.length).to eq(Sentry::Event::MAX_MESSAGE_SIZE_IN_BYTES + 1)
+    end
+  end
+
   describe "#to_hash" do
     it "serializes data correctly" do
       result = crumb.to_hash