+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Fixes the issue mentioned in this [comment](https://github.com/getsentry/sentry-ruby/pull/1199#issuecomment-773069840)
 - Correct the timing of loading ActiveJobExtensions [#1464](https://github.com/getsentry/sentry-ruby/pull/1464)
   - Fixes [#1249](https://github.com/getsentry/sentry-ruby/issues/1249)
+- Limit breadcrumb's message length [#1465](https://github.com/getsentry/sentry-ruby/pull/1465)
 
 ## 4.5.0